add a flush_output method to HostOutputStream and WasiCtx.
When a Table is dropped after component execution, it will also drop any
HostOutputStreams created inside it. This may happen before those
streams have finished sending any buffered output. The `flush_output`
method is a special way to destroy a WasiCtx which will remove all of
the HostOutputStreams from the table and await for them to flush their
outputs.